agent/requirements.txt
langchain
langchain-openai
langchain-google-genai
langchain-community
python-dotenv
requests
tiktoken
opencv-python
numpy
google-generativeai
browser-use
nest-asyncio
playwright